refactor: feature 별로 패키지 관리
