+++ src/main/java/subway/station/StationRepository.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.station;
 
 import java.util.ArrayList;
 import java.util.Collections;
